% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suminagashi.R
\name{scale_fill_suminagashi}
\alias{scale_fill_suminagashi}
\title{Define colour palette for elevation map}
\usage{
scale_fill_suminagashi(breaks, pal = NULL)
}
\arguments{
\item{breaks}{Breaks in contours}

\item{pal}{Colour palette}
}
\description{
Define colour palette for elevation map
}
