% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suminagashi.R
\name{suminagashi}
\alias{suminagashi}
\title{Make suminagashi-style topographic map art}
\usage{
suminagashi(elev, xlim, ylim, pal_breaks, pal, file_name, bd = FALSE)
}
\arguments{
\item{elev}{An elevation raster object of the mapping area}

\item{xlim}{Limits for x-axis, format c(xmin, xmax)}

\item{ylim}{Limits for y-axis, format c(ymin, ymax)}

\item{pal_breaks}{Breaks in contours for palette}

\item{pal}{Colour palette for contours; repeated if shorter than \code{pal_breaks}}

\item{file_name}{Optional: file name and extension}
}
\description{
Make suminagashi-style topographic map art
}
